Allow larger costume quantities

+++ costume_store.sql
@@ -29,7 +29,7 @@ CREATE TABLE CostumeStore.dbo.CostumeSales
     CustomerLastName    VARCHAR(20) NOT NULL,
     CostumeName         VARCHAR(25) NOT NULL,
     Size                CHAR(2) NOT NULL CHECK (Size IN ('XS','S','M','L','XL')),
-    Quantity            TINYINT NOT NULL CHECK (Quantity > 0),
+    Quantity            SMALLINT NOT NULL CHECK (Quantity > 0),
     SoldPricePerCostume DECIMAL(5,2) NOT NULL CHECK (SoldPricePerCostume >= 0),
     DateBought          DATE NOT NULL,
     DateSold            DATE NOT NULL,
